+++ machine_learning_hep/doskimming.py
@@ -21,79 +21,88 @@ main macro for charm analysis with python
 
 # pylint: disable=import-error
 
-import subprocess
-import os
 import multiprocessing as mp
 import uproot
+import pickle
+import pandas as pd
 from machine_learning_hep.general import get_database_ml_parameters
-from machine_learning_hep.root import write_tree
-from machine_learning_hep.listfiles import list_files
-
-def flattenroot_to_pandas(namefileinput, namefileoutput, treenamein, \
-                          treenameout, var_all, skimming_sel, writeflat):
-    tree = uproot.open(namefileinput)[treenamein]
-    df = tree.pandas.df(branches=var_all, flatten=True)
-    df = df.query(skimming_sel)
-    df.to_pickle(namefileoutput)
-    namefileoutputrootflat = namefileoutput.replace(".pkl", "Flat.root")
-    if writeflat is True:
-        write_tree(namefileoutputrootflat, treenameout, df)
+from machine_learning_hep.listfiles import list_files_dir
 
 def writelist_tofile(fileout, mylist):
     with open(fileout, 'w') as f:
         for item in mylist:
             f.write("%s\n" % item)
 
-def merger_skimmer(listfiles, index, directory, case, treenamein, \
-                   treenameout, var_all_unflat, skimming_sel, writeflat, \
-                   writenasted, nmaxfile):
-    filemerged = "%s/AnalysisResultsMergedIndex%d_Nfiles%d.root" % (directory, index, nmaxfile)
-    if not os.path.isfile(filemerged):
-        filelist = "%s/mergedfilesMergedIndex%d_Nfiles%d.txt" % (directory, index, nmaxfile)
-        writelist_tofile(filelist, listfiles)
-        subprocess.call("hadd %s @%s" % (filemerged, filelist), shell=True)
-    filepandas = filemerged.replace('.root', '%s.pkl' % (case))
-    flattenroot_to_pandas(filemerged, filepandas, treenamein, treenameout, \
-                          var_all_unflat, skimming_sel, writeflat)
-    if writenasted is False:
-        os.remove(filemerged)
+def flattenroot_to_pandas(filein, fileout, treenamein, var_all, skimming_sel):
+    print(filein)
+    tree = uproot.open(filein)[treenamein]
+    df = tree.pandas.df(branches=var_all, flatten=True)
+    df = df.query(skimming_sel)
+    df.to_pickle(fileout)
 
-def mergefiles(chunks, mergeddir, case, treenamein, treenameout, \
-               var_all_unflat, skimming_sel, writeflat, writenasted, nmaxfile):
-    processes = [mp.Process(target=merger_skimmer, args=(mylist, index, mergeddir, case, \
-                                                         treenamein, treenameout, var_all_unflat, \
-                                                         skimming_sel, \
-                                                         writeflat, writenasted, nmaxfile))
-                 for index, mylist in enumerate(chunks)]
+def flattenallpickle(chunk, chunkout, treenamein, var_all, skimming_sel):
+    processes = [mp.Process(target=flattenroot_to_pandas, args=(filein, chunkout[index], \
+                                                         treenamein, var_all, skimming_sel))
+                 for index, filein in enumerate(chunk)]
+    for p in processes:
+        p.start()
+    for p in processes:
+        p.join()
+
+def merge(chunk, mergeddir, index, case):
+    dfList = []
+    for myfilename in chunk:
+        myfile = open (myfilename, "rb")
+        df = pickle.load(myfile)
+        dfList.append(df)
+    dftot = pd.concat(dfList)
+    namemerged = "%s/AnalysisResults%sMergedIndex%d.pkl" % (mergeddir, case, index)
+    dftot.to_pickle(namemerged)
+
+def mergeall(chunksmerge, mergeddir, case):
+    processes = [mp.Process(target=merge, args=(chunk, mergeddir, index, case))
+                 for index, chunk in enumerate(chunksmerge)]
     for p in processes:
         p.start()
     for p in processes:
         p.join()
 
 # pylint: disable=too-many-locals, too-many-statements, too-many-branches
-def doskimming():
+def doskimming2():
 
     case = "Dzero"
     namefileinput = 'AnalysisResults.root'
+    namefileinputpkl = 'AnalysisResults%s.pkl' % case
     treenamein = 'PWGHF_TreeCreator/tree_D0'
-    treenameout = 'tree_D0flat'
     data = get_database_ml_parameters()
     var_all_unflat = data[case]["var_all_unflat"]
-    skimming_sel = "n_cand> 0 & pt_cand>3"
-    nmaxfile = 5
-    nmaxfilestoprocess = 100
-    writeflat = True
-    writenasted = False
+    skimming_sel = "n_cand> 0 & pt_cand>2"
+    nmaxchunks = 50
+    nmaxfiles = 50
+    nmaxmerge = 10
 
-#    inputdir = "/home/ginnocen/LearningPythonML/inputs"
+    doconversion = 1
+    domerge = 1
+
+    #inputdir = "/home/ginnocen/LearningPythonML/inputs"
     inputdir = "/data/HeavyFlavour/DmesonsLc_pp_5TeV/Data_LHC17pq/12-02-2019/340_20190211-2126/unmerged" # pylint: disable=line-too-long
-    mergeddir = "/data/HeavyFlavour/DmesonsLc_pp_5TeV/Data_LHC17pq/12-02-2019/340_20190211-2126/mergedPython" # pylint: disable=line-too-long
+    mergeddir = \
+    "/data/HeavyFlavour/DmesonsLc_pp_5TeV/Data_LHC17pq/12-02-2019/340_20190211-2126/unmergedpkl" # pylint: disable=line-too-long
 
-    listfilespath, _ = list_files(inputdir, outdir="", \
-                                  filenameinput=namefileinput, filenameoutput="")
-    listfilespath = listfilespath[:nmaxfilestoprocess]
-    chunks = [listfilespath[x:x+nmaxfile] for x in range(0, len(listfilespath), nmaxfile)]
-    mergefiles(chunks, mergeddir, case, treenamein, treenameout,
-               var_all_unflat, skimming_sel, writeflat, writenasted, nmaxfile)
+    listfilespath, listfilespathout = list_files_dir(inputdir, outdir=mergeddir, \
+                                  filenameinput=namefileinput, filenameoutput=namefileinputpkl)
+    listfilespath = listfilespath[:nmaxfiles]
+    listfilespathout = listfilespathout[:nmaxfiles]
+    print(len(listfilespath))
+    print(len(listfilespathout))
+    if doconversion == 1:
+        chunks = [listfilespath[x:x+nmaxchunks] for x in range(0, len(listfilespath), nmaxchunks)]
+        chunksout = [listfilespathout[x:x+nmaxchunks] for x in range(0, len(listfilespathout), nmaxchunks)]
+        for chunk, chunkout in zip(chunks, chunksout):
+            flattenallpickle(chunk, chunkout, treenamein, var_all_unflat, skimming_sel)
+    if domerge == 1:
+        chunksmerge = [listfilespathout[x:x+nmaxmerge] \
+                   for x in range(0, len(listfilespathout), nmaxmerge)]
+        mergeall(chunksmerge, mergeddir,  case)
 
-doskimming()
+doskimming2()
